follow-up: unify drop namespace error
